pyproject.toml
[project]
name = "hackseoul-fe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.0",
    "google-genai>=1.49.0",
    "openai>=1.54.0",
    "anthropic>=0.34.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "aiohttp>=3.9.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
]
